[BUGFIX] Wrong XPath namespace registered when parsing multistatus messages
The namespaces is converted from DAV: to urn:DAV: in
Sabre_DAV_XMLUtil::convertDAVNamespace() to comply with XML namespacing
rules. This is however not reflected in
Sabre_DAV_Client::parseMultiStatus(), leading to an error if anything
else than d: is used by the server as the namespace prefix in DAV
responses (e.g. Apache 2.2 uses D:, which leads to strange behaviour
because the XML is case-sensitive).