[CalDAV] Fixed validateICalendar() to keep data
The `validateICalendar()` method gets content of a `$data` parameter,
if this parameter is a resource, then it uses `stream_get_contents()`
to check real data.

But it's wrong as the resource has been consumed, and it won't keep
data for next steps (in the `put()` method for instance).

So, to pass the `$data` parameter as reference seems fine.