A new approach to method handling.
Every method handler now gets a reference to the request and response,
making it easier to do sub-requests.

Standard method handlers are moving to a core plugin.

It's now possible to listen for specific methods, such as
beforeMethod:PATCH.

New event: 'method' for a standard method handler.
New event: 'afterMethod' for post-processing

Server is now responsible for calling $response->send();