Improved AbstractBasic.php log messages
There are two possible log messages in function authenticate.

a.) The current message text "No basic authentication headers were
found" it is not clear what is going on and if there is a problem or
not. According to a explanation of @evert, this can also be triggered
when a client (eg a browser) tries to login and with this first attempt
no headers are sent. Then they negotiate and with the second attempt the
authentication method negotiated used is clear. Therefore I adopted this message to make this negotiating process more visible and better understandable.

b.) The second message just tells that there is a mismatch between a
user and a password given. Currently no information is logged with which user this attempt was made. I extended the message text to printout the user used making it easier finding login related issues.