Fix negated text-match on param-filter
The negation applies to the individual text-match (i.e. the negated
text-match matches if a parameter string does not match the search
string). It does not apply to the entire param-filter (i.e. it does not
mean there must be NO property instance that matches the param filter,
it is sufficient if there is one that does not match).