Fix showing owner of shared calendars on iOS
The iOS calendar uses the {DAV:}owner attribute to determine the owner
of a shared calendar.

Therefore this attribute needs to be set as the real owner of the
calendar not the user which whom the calendar is shared.
This "real" owner needs to be set before the ACL plugin can set it.

To determine the name of the owner the iOS calendar queries the principal uri.
Setting a mailto address is not working.

Unfortunatly the principal query is not allowed for other users by the
ACL plugin. It is not easy to determine if the client is trying to
query the principal information of a owner of a shared calendar, so as
workaround all authenticated users are allowed to query the principal information.