Uri in xml response must be url encoded.
The Uri is derived from the request with the method getPath, which generates a valid path fragment suitable for file and directory names.

In the XML-response to a lock-request the uri is used for lockroot element. If the uri contains a space or german umlauts, a strict client like libneon, which is used in davfs2, refuses the lock reponse and files cannot be copied.